requirements.txt
PyYAML
colorama
http-status
nose2
pyparsing
pyquery
requests
wolframalpha


git+https://github.com/Rapptz/discord.py.git@rewrite
